+++ src/Countries/BelizeCountryConfig.php
@@ -12,9 +12,9 @@ abstract class BelizeCountryConfig extends CountryConfig
         return Country::BZ === $countryCode;
     }
 
-    public static function getEndpoints(): array
+    public static function getEndpoints(string $client): array
     {
-        return array_merge(parent::getEndpoints(), [
+        return array_merge(parent::getEndpoints($client), [
             Environment::PROD => 'https://abgateway.atlabank.com',
         ]);
     }