+++ src/Countries/HondurasCountryConfig.php
@@ -12,9 +12,9 @@ abstract class HondurasCountryConfig extends CountryConfig
         return Country::HN === $countryCode;
     }
 
-    public static function getEndpoints(): array
+    public static function getEndpoints(string $client): array
     {
-        return array_merge(parent::getEndpoints(), [
+        return array_merge(parent::getEndpoints($client), [
             Environment::PROD => 'https://pagoenlinea.bancatlan.hn',
         ]);
     }